# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface ContractExtractionUpload {
  /** A PDF contract up to 10 MB. */
  file: Blob;
}

export type ContractValueStatus = typeof ContractValueStatus[keyof typeof ContractValueStatus];


export const ContractValueStatus = {
  stated: 'stated',
  unknown: 'unknown',
} as const;

export interface ContractValue {
  status: ContractValueStatus;
  /** @nullable */
  amount?: number | null;
  /** @nullable */
  currency?: string | null;
}

export type ContractConfidenceVendor = typeof ContractConfidenceVendor[keyof typeof ContractConfidenceVendor];


export const ContractConfidenceVendor = {
  High: 'High',
  Medium: 'Medium',
  Low: 'Low',
} as const;

export type ContractConfidenceContractNumber = typeof ContractConfidenceContractNumber[keyof typeof ContractConfidenceContractNumber];


export const ContractConfidenceContractNumber = {
  High: 'High',
  Medium: 'Medium',
  Low: 'Low',
} as const;

export type ContractConfidenceContractName = typeof ContractConfidenceContractName[keyof typeof ContractConfidenceContractName];


export const ContractConfidenceContractName = {
  High: 'High',
  Medium: 'Medium',
  Low: 'Low',
} as const;

export type ContractConfidenceContractType = typeof ContractConfidenceContractType[keyof typeof ContractConfidenceContractType];


export const ContractConfidenceContractType = {
  High: 'High',
  Medium: 'Medium',
  Low: 'Low',
} as const;

export type ContractConfidenceContractValue = typeof ContractConfidenceContractValue[keyof typeof ContractConfidenceContractValue];


export const ContractConfidenceContractValue = {
  High: 'High',
  Medium: 'Medium',
  Low: 'Low',
} as const;

export type ContractConfidenceStartDate = typeof ContractConfidenceStartDate[keyof typeof ContractConfidenceStartDate];


export const ContractConfidenceStartDate = {
  High: 'High',
  Medium: 'Medium',
  Low: 'Low',
} as const;

export type ContractConfidenceContractDuration = typeof ContractConfidenceContractDuration[keyof typeof ContractConfidenceContractDuration];


export const ContractConfidenceContractDuration = {
  High: 'High',
  Medium: 'Medium',
  Low: 'Low',
} as const;

export type ContractConfidenceEndDate = typeof ContractConfidenceEndDate[keyof typeof ContractConfidenceEndDate];


export const ContractConfidenceEndDate = {
  High: 'High',
  Medium: 'Medium',
  Low: 'Low',
} as const;

export type ContractConfidenceNoticePeriod = typeof ContractConfidenceNoticePeriod[keyof typeof ContractConfidenceNoticePeriod];


export const ContractConfidenceNoticePeriod = {
  High: 'High',
  Medium: 'Medium',
  Low: 'Low',
} as const;

export type ContractConfidenceNoticeDeadline = typeof ContractConfidenceNoticeDeadline[keyof typeof ContractConfidenceNoticeDeadline];


export const ContractConfidenceNoticeDeadline = {
  High: 'High',
  Medium: 'Medium',
  Low: 'Low',
} as const;

export type ContractConfidenceNegotiationBuffer = typeof ContractConfidenceNegotiationBuffer[keyof typeof ContractConfidenceNegotiationBuffer];


export const ContractConfidenceNegotiationBuffer = {
  High: 'High',
  Medium: 'Medium',
  Low: 'Low',
} as const;

export type ContractConfidenceOwner = typeof ContractConfidenceOwner[keyof typeof ContractConfidenceOwner];


export const ContractConfidenceOwner = {
  High: 'High',
  Medium: 'Medium',
  Low: 'Low',
} as const;

export type ContractConfidenceStatus = typeof ContractConfidenceStatus[keyof typeof ContractConfidenceStatus];


export const ContractConfidenceStatus = {
  High: 'High',
  Medium: 'Medium',
  Low: 'Low',
} as const;

export interface ContractConfidence {
  vendor: ContractConfidenceVendor;
  contractNumber: ContractConfidenceContractNumber;
  contractName: ContractConfidenceContractName;
  contractType: ContractConfidenceContractType;
  contractValue: ContractConfidenceContractValue;
  startDate: ContractConfidenceStartDate;
  contractDuration: ContractConfidenceContractDuration;
  endDate: ContractConfidenceEndDate;
  noticePeriod: ContractConfidenceNoticePeriod;
  noticeDeadline: ContractConfidenceNoticeDeadline;
  negotiationBuffer: ContractConfidenceNegotiationBuffer;
  owner: ContractConfidenceOwner;
  status: ContractConfidenceStatus;
}

export type ContractReviewDraftContractType = typeof ContractReviewDraftContractType[keyof typeof ContractReviewDraftContractType];


export const ContractReviewDraftContractType = {
  '': '',
  Maintenance: 'Maintenance',
  Software_License: 'Software License',
  Real_Estate: 'Real Estate',
  Infrastructure: 'Infrastructure',
} as const;

export type ContractReviewDraftStatus = typeof ContractReviewDraftStatus[keyof typeof ContractReviewDraftStatus];


export const ContractReviewDraftStatus = {
  At_Risk: 'At Risk',
  Review_Open: 'Review Open',
  In_Negotiation: 'In Negotiation',
} as const;

export interface ContractReviewDraft {
  vendor: string;
  contractNumber: string;
  contractName: string;
  contractType: ContractReviewDraftContractType;
  contractValue: ContractValue;
  startDate: string;
  contractDuration: string;
  endDate: string;
  noticePeriod: string;
  noticeDeadline: string;
  negotiationBuffer: string;
  owner: string;
  status: ContractReviewDraftStatus;
}

export type ContractExtractionResultExtraction = {
  contract: ContractReviewDraft;
  confidence: ContractConfidence;
};

export interface ContractExtractionResult {
  filename: string;
  extraction: ContractExtractionResultExtraction;
}