# lib/api-zod/src/generated/types/contractExtractionResult.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ContractExtractionResultExtraction } from './contractExtractionResultExtraction';

export interface ContractExtractionResult {
  filename: string;
  extraction: ContractExtractionResultExtraction;
}