# lib/api-zod/src/generated/types/contractReviewDraft.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ContractReviewDraftContractType } from './contractReviewDraftContractType';
import type { ContractReviewDraftStatus } from './contractReviewDraftStatus';
import type { ContractValue } from './contractValue';

export interface ContractReviewDraft {
  vendor: string;
  contractNumber: string;
  contractName: string;
  contractType: ContractReviewDraftContractType;
  contractValue: ContractValue;
  startDate: string;
  contractDuration: string;
  endDate: string;
  noticePeriod: string;
  noticeDeadline: string;
  negotiationBuffer: string;
  owner: string;
  status: ContractReviewDraftStatus;
}