# lib/api-zod/src/generated/types/contractValue.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ContractValueStatus } from './contractValueStatus';

export interface ContractValue {
  status: ContractValueStatus;
  /** @nullable */
  amount?: number | null;
  /** @nullable */
  currency?: string | null;
}